pyproject.toml
[project]
name = "web-mining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "streamlit>=1.32.0",
    "pandas>=2.2.0",
    "networkx>=3.2",
    "pyvis>=0.3.2",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]
